refactor(spa): extract 4 modal components from AppComponent (T4e) [KAN-118]
Extract AuthModal, CreateCookbookModal, ManualEntryModal, and
AddToCookbookModal into self-contained components under src/modals/.
AppComponent drops from 469→119 lines TS, 715→85 lines HTML.
Cross-modal create-from-add-flow coordinated via viewChild + events.
Tests migrated to target new components directly.
