+++ src/app.component.test.ts
@@ -1,17 +1,16 @@
 import '@angular/compiler';
 import { Injector, runInInjectionContext, signal } from '@angular/core';
 import { Router } from '@angular/router';
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { AppComponent } from './app.component';
+import { describe, expect, it, vi } from 'vitest';
 import { GeneratorComponent } from './components/generator/generator.component';
+import { ManualEntryModalComponent } from './modals/manual-entry/manual-entry-modal.component';
+import { CreateCookbookModalComponent } from './modals/create-cookbook/create-cookbook-modal.component';
 import { AuthService } from './services/auth.service';
 import { GeminiService } from './services/gemini.service';
 import { PersistenceService } from './services/persistence.service';
-import { ToastService } from './services/toast.service';
 import { RecipeStateService } from './services/recipe-state.service';
 
 const mockRouter = { events: { subscribe: vi.fn() }, navigate: vi.fn() };
-const mockToastService = { toasts: signal([]), show: vi.fn(), dismiss: vi.fn() };
 const mockRecipeStateService = {
   currentRecipe: signal(null),
   generatedImageUrl: signal(null),
@@ -20,46 +19,14 @@ const mockRecipeStateService = {
   clearRecipe: vi.fn(),
 };
 
-describe('AppComponent manual recipe entry', () => {
-  beforeEach(() => {
-    vi.stubGlobal('window', {
-      addEventListener: vi.fn(),
-      location: {
-        hash: '',
-        href: 'http://localhost/',
-        pathname: '/',
-        search: '',
-      },
-      history: {
-        replaceState: vi.fn(),
-      },
-    });
-  });
-
-  afterEach(() => {
-    vi.unstubAllGlobals();
-  });
-
+describe('ManualEntryModalComponent reopened session resets drafts', () => {
   it('starts a reopened manual-entry session with empty ingredient and instruction drafts', () => {
     const injector = Injector.create({
-      providers: [
-        { provide: Router, useValue: mockRouter },
-        { provide: ToastService, useValue: mockToastService },
-        { provide: RecipeStateService, useValue: mockRecipeStateService },
-        { provide: GeminiService, useValue: {} },
-        { provide: PersistenceService, useValue: {} },
-        {
-          provide: AuthService,
-          useValue: {
-            ensureGuestSession: vi.fn(),
-            ready: Promise.resolve(),
-          },
-        },
-      ],
+      providers: [{ provide: PersistenceService, useValue: {} }],
     });
-    const component = runInInjectionContext(injector, () => new AppComponent());
+    const component = runInInjectionContext(injector, () => new ManualEntryModalComponent());
 
-    component.openManualEntryModal();
+    component.open();
     component.newIngredient.set({
       name: 'Stale tofu',
       amount: 2,
@@ -67,9 +34,9 @@ describe('AppComponent manual recipe entry', () => {
       type: 'wet',
     });
     component.newInstruction.set('Carry this into the next recipe');
-    component.closeManualEntryModal();
+    component.close();
 
-    component.openManualEntryModal();
+    component.open();
 
     expect(component.newIngredient()).toEqual({
       name: '',
@@ -102,17 +69,6 @@ describe('GeneratorComponent slugFromTitle parity with Backend normalize_slug', 
     return runInInjectionContext(injector, () => new GeneratorComponent());
   };
 
-  // Expected values are the literal output of the server implementation
-  // (Backend/utils/slug_utils.normalize_slug), the authority this method
-  // mirrors. Regenerate after changing either side:
-  //   cd Backend && uv run python -c \
-  //     "from utils.slug_utils import normalize_slug; print(normalize_slug('<title>'))"
-  // Probes chosen to catch known divergence classes: characters NFKD cannot
-  // decompose (ø, Ł, ß, þ, ð — deleted by the server, so no hyphen may appear
-  // in their place), decomposable diacritics (é, ü, ï), non-ASCII punctuation
-  // (em/en dashes — deleted, NOT collapsed to '-'), and inputs with no ASCII
-  // alphanumerics at all (CJK, emoji), which yield '' and are rejected
-  // server-side with 400 PUBLIC_SLUG_REQUIRED_ERROR on publish.
   const parityCases: Array<[title: string, expected: string]> = [
     ["Adam's Amazing Pasta", 'adam-s-amazing-pasta'],
     ['Café Crème Brûlée', 'cafe-creme-brulee'],
@@ -137,44 +93,14 @@ describe('GeneratorComponent slugFromTitle parity with Backend normalize_slug', 
   });
 });
 
-describe('AppComponent.createCookbook in-flight guard', () => {
-  beforeEach(() => {
-    vi.stubGlobal('window', {
-      addEventListener: vi.fn(),
-      location: {
-        hash: '',
-        href: 'http://localhost/',
-        pathname: '/',
-        search: '',
-      },
-      history: {
-        replaceState: vi.fn(),
-      },
-    });
-  });
-
-  afterEach(() => {
-    vi.unstubAllGlobals();
-  });
-
+describe('CreateCookbookModalComponent in-flight guard', () => {
   const createComponent = (createCookbookImpl: (...args: unknown[]) => Promise<string | null>) => {
     const injector = Injector.create({
       providers: [
-        { provide: Router, useValue: mockRouter },
-        { provide: ToastService, useValue: mockToastService },
-        { provide: RecipeStateService, useValue: mockRecipeStateService },
-        { provide: GeminiService, useValue: {} },
         { provide: PersistenceService, useValue: { createCookbook: vi.fn(createCookbookImpl) } },
-        {
-          provide: AuthService,
-          useValue: {
-            ensureGuestSession: vi.fn(),
-            ready: Promise.resolve(),
-          },
-        },
       ],
     });
-    return runInInjectionContext(injector, () => new AppComponent());
+    return runInInjectionContext(injector, () => new CreateCookbookModalComponent());
   };
 
   it('ignores a second call while the first request is still in flight', async () => {
@@ -185,9 +111,9 @@ describe('AppComponent.createCookbook in-flight guard', () => {
     const component = createComponent(() => inFlight);
     component.newCookbookName.set('Weeknight Dinners');
 
-    const first = component.createCookbook();
+    const first = component.create();
     expect(component.isCreatingCookbook()).toBe(true);
-    const second = component.createCookbook(); // should short-circuit, not fire a second request
+    const second = component.create();
 
     resolveCreate('cookbook-1');
     await Promise.all([first, second]);
@@ -196,29 +122,34 @@ describe('AppComponent.createCookbook in-flight guard', () => {
       (component['persistenceService'].createCookbook as ReturnType<typeof vi.fn>).mock.calls
     ).toHaveLength(1);
     expect(component.isCreatingCookbook()).toBe(false);
-    expect(component.showCreateCookbookModal()).toBe(false);
+    expect(component.isOpen()).toBe(false);
   });
 
   it('resets the guard and keeps the modal open when creation is rejected', async () => {
-    const component = createComponent(async () => null); // e.g. a 400/401/403 rejection
+    const component = createComponent(async () => null);
     component.newCookbookName.set('Weeknight Dinners');
-    component.showCreateCookbookModal.set(true);
+    component.isOpen.set(true);
 
-    await component.createCookbook();
+    await component.create();
 
     expect(component.isCreatingCookbook()).toBe(false);
-    expect(component.showCreateCookbookModal()).toBe(true);
+    expect(component.isOpen()).toBe(true);
   });
 
-  it('adds the server-resolved cookbook id, not a guessed one, to pendingCookbookIds', async () => {
+  it('emits cookbookCreated with the server-resolved id', async () => {
     const component = createComponent(async () => 'resolved-id');
     component.newCookbookName.set('Weeknight Dinners');
-    component['_creatingFromAddFlow'] = true;
+    component.isOpen.set(true);
 
-    await component.createCookbook();
+    let emittedId: string | undefined;
+    component.cookbookCreated.subscribe((id) => {
+      emittedId = id;
+    });
 
-    expect(component.pendingCookbookIds().has('resolved-id')).toBe(true);
-    expect(component.showCreateCookbookModal()).toBe(false);
+    await component.create();
+
+    expect(emittedId).toBe('resolved-id');
+    expect(component.isOpen()).toBe(false);
   });
 });
