+++ src/modals/add-to-cookbook/add-to-cookbook-modal.component.html
@@ -1,0 +1,103 @@
+@if (isOpen()) {
+  <div class="fixed inset-0 z-50 flex items-center justify-center p-4">
+    <button
+      type="button"
+      class="absolute inset-0 bg-stone-900/50 backdrop-blur-sm"
+      (click)="close()"
+      aria-label="Close cookbook selection dialog"
+      tabindex="-1"
+    ></button>
+    <div
+      class="relative bg-white w-full max-w-sm rounded-3xl shadow-2xl p-6 animate-[fadeIn_0.3s_ease-out]"
+    >
+      <div class="flex justify-between items-center mb-4">
+        <h3 class="text-xl font-bold text-stone-800 serif">Add to Cookbook</h3>
+        <button
+          (click)="createCookbookRequested.emit()"
+          class="text-sm text-green-700 font-bold hover:underline"
+        >
+          + New
+        </button>
+      </div>
+
+      <div class="space-y-2 max-h-[300px] overflow-y-auto mb-6">
+        @if (authService.currentUser()?.cookbooks?.length === 0) {
+          <p class="text-stone-400 text-sm text-center py-4">No cookbooks yet. Create one!</p>
+        }
+
+        @for (cb of authService.currentUser()?.cookbooks; track cb.id) {
+          <button
+            (click)="toggleCookbookSelection(cb.id)"
+            class="w-full text-left px-4 py-3 rounded-xl transition-colors flex justify-between items-center group"
+            [class.bg-green-50]="isCookbookSelected(cb.id)"
+            [class.bg-stone-50]="!isCookbookSelected(cb.id)"
+            [class.ring-2]="isCookbookSelected(cb.id)"
+            [class.ring-green-500]="isCookbookSelected(cb.id)"
+          >
+            <span
+              class="font-medium"
+              [class.text-green-800]="isCookbookSelected(cb.id)"
+              [class.text-stone-700]="!isCookbookSelected(cb.id)"
+              >{{ cb.name }}</span
+            >
+            @if (isCookbookSelected(cb.id)) {
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                class="h-5 w-5 text-green-600"
+                viewBox="0 0 20 20"
+                fill="currentColor"
+              >
+                <path
+                  fill-rule="evenodd"
+                  d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z"
+                  clip-rule="evenodd"
+                />
+              </svg>
+            }
+          </button>
+        }
+      </div>
+
+      @if (showRemoveConfirmation()) {
+        <div class="mb-4 p-3 bg-amber-50 border border-amber-200 rounded-xl text-sm">
+          <p class="font-bold text-amber-800 mb-1">Remove from cookbooks?</p>
+          <p class="text-amber-700">
+            This recipe will be removed from:
+            <strong>{{ cookbooksBeingRemoved().join(', ') }}</strong>
+          </p>
+          <div class="flex gap-2 mt-3">
+            <button
+              (click)="cancelRemoveConfirmation()"
+              class="flex-1 py-2 bg-stone-100 text-stone-600 font-bold rounded-lg hover:bg-stone-200 text-xs"
+            >
+              Go Back
+            </button>
+            <button
+              (click)="applyCookbookChanges()"
+              class="flex-1 py-2 bg-amber-600 text-white font-bold rounded-lg hover:bg-amber-700 text-xs"
+            >
+              Yes, Remove
+            </button>
+          </div>
+        </div>
+      }
+
+      <div class="flex gap-3">
+        <button
+          (click)="close()"
+          class="flex-1 py-3 bg-stone-100 text-stone-600 font-bold rounded-xl hover:bg-stone-200"
+        >
+          Cancel
+        </button>
+        @if (!showRemoveConfirmation()) {
+          <button
+            (click)="confirmCookbookChanges()"
+            class="flex-1 py-3 bg-green-600 text-white font-bold rounded-xl hover:bg-green-700 shadow-lg shadow-green-600/20"
+          >
+            Confirm
+          </button>
+        }
+      </div>
+    </div>
+  </div>
+}
