+++ src/modals/auth/auth-modal.component.html
@@ -1,0 +1,155 @@
+@if (isOpen()) {
+  <div class="fixed inset-0 z-50 flex items-center justify-center p-4">
+    <button
+      type="button"
+      class="absolute inset-0 bg-stone-900/50 backdrop-blur-sm"
+      (click)="close()"
+      aria-label="Close sign-in dialog"
+      tabindex="-1"
+    ></button>
+    <div
+      class="relative bg-white w-full max-w-md rounded-3xl shadow-2xl overflow-hidden animate-[fadeIn_0.3s_ease-out]"
+    >
+      <button
+        (click)="close()"
+        aria-label="Close sign-in dialog"
+        class="absolute top-4 right-4 text-stone-400 hover:text-stone-600"
+      >
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          class="h-6 w-6"
+          fill="none"
+          viewBox="0 0 24 24"
+          stroke="currentColor"
+        >
+          <path
+            stroke-linecap="round"
+            stroke-linejoin="round"
+            stroke-width="2"
+            d="M6 18L18 6M6 6l12 12"
+          />
+        </svg>
+      </button>
+
+      <div class="p-8">
+        <div class="text-center mb-8">
+          <div class="inline-flex p-3 rounded-full bg-green-100 text-green-700 mb-4">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              class="h-8 w-8"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+            >
+              <path
+                stroke-linecap="round"
+                stroke-linejoin="round"
+                stroke-width="2"
+                d="M3.055 11H5a2 2 0 012 2v1a2 2 0 002 2 2 2 0 012 2v2.945M8 3.935V5.5A2.5 2.5 0 0010.5 8h.5a2 2 0 012 2 2 2 0 104 0 2 2 0 012-2h1.064M15 20.488V18a2 2 0 012-2h3.064M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
+              />
+            </svg>
+          </div>
+          <h2 class="text-3xl font-bold text-stone-800 mb-2 serif">Welcome to VeganGenius</h2>
+          <p class="text-stone-500">
+            Sign in to save recipes, create cookbooks, and sync your kitchen across devices.
+          </p>
+        </div>
+
+        @if (authError()) {
+          <div
+            class="mb-4 p-3 bg-red-50 text-red-700 text-sm rounded-lg border border-red-200 text-center"
+          >
+            {{ authError() }}
+          </div>
+        }
+
+        @if (isInAppBrowser()) {
+          <div class="rounded-xl border-2 border-amber-200 bg-amber-50 p-4 text-left">
+            <p class="text-sm font-semibold text-amber-900 mb-1">Finish sign-in in your browser</p>
+            <p class="text-sm text-amber-800">
+              You're viewing this inside another app, and Google won't let you sign in here. Open
+              this page in Safari or Chrome to sign in with Google.
+            </p>
+            <button
+              (click)="copyShareLink()"
+              class="mt-3 w-full flex items-center justify-center gap-2 px-4 py-3 bg-white border-2 border-amber-300 rounded-lg text-amber-900 font-semibold hover:bg-amber-100 transition-colors duration-200"
+            >
+              @if (copiedShareLink()) {
+                <span>Link copied — paste it in your browser</span>
+              } @else {
+                <span>Copy page link</span>
+              }
+            </button>
+            <p class="mt-2 text-xs text-amber-700">
+              Tip: tap the "⋯" or share menu, then "Open in Safari" / "Open in Chrome".
+            </p>
+          </div>
+        } @else {
+          <button
+            (click)="onGoogleLogin()"
+            [disabled]="authLoginLoading()"
+            class="w-full flex items-center justify-center gap-3 px-6 py-4 bg-white border-2 border-stone-200 rounded-xl hover:bg-stone-50 hover:border-stone-300 disabled:opacity-70 transition-all duration-200 shadow-sm"
+          >
+            @if (authLoginLoading()) {
+              <svg
+                class="animate-spin h-5 w-5 text-stone-500"
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 24 24"
+              >
+                <circle
+                  class="opacity-25"
+                  cx="12"
+                  cy="12"
+                  r="10"
+                  stroke="currentColor"
+                  stroke-width="4"
+                ></circle>
+                <path
+                  class="opacity-75"
+                  fill="currentColor"
+                  d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"
+                ></path>
+              </svg>
+              <span class="text-stone-600 font-semibold">Connecting...</span>
+            } @else {
+              <svg class="h-5 w-5" viewBox="0 0 24 24" fill="currentColor">
+                <path
+                  d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
+                  fill="#4285F4"
+                />
+                <path
+                  d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
+                  fill="#34A853"
+                />
+                <path
+                  d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
+                  fill="#FBBC05"
+                />
+                <path
+                  d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
+                  fill="#EA4335"
+                />
+              </svg>
+              <span class="text-stone-700 font-semibold">Sign in with Google</span>
+            }
+          </button>
+        }
+
+        <div class="relative my-6">
+          <div class="absolute inset-0 flex items-center">
+            <div class="w-full border-t border-stone-200"></div>
+          </div>
+          <div class="relative flex justify-center text-sm">
+            <span class="px-2 bg-white text-stone-400">or</span>
+          </div>
+        </div>
+
+        <p class="text-center text-sm text-stone-500">
+          No account needed to generate recipes! Close this dialog to continue as a guest. Your
+          recipes will be stored locally in this browser.
+        </p>
+      </div>
+    </div>
+  </div>
+}
