+++ src/modals/create-cookbook/create-cookbook-modal.component.html
@@ -1,0 +1,43 @@
+@if (isOpen()) {
+  <div class="fixed inset-0 z-[60] flex items-center justify-center p-4">
+    <button
+      type="button"
+      class="absolute inset-0 bg-stone-900/50 backdrop-blur-sm"
+      (click)="close()"
+      aria-label="Close cookbook dialog"
+      tabindex="-1"
+    ></button>
+    <div
+      class="relative bg-white w-full max-w-sm rounded-3xl shadow-2xl p-6 animate-[fadeIn_0.3s_ease-out]"
+    >
+      <h3 class="text-xl font-bold text-stone-800 mb-4 serif">New Cookbook</h3>
+      <input
+        type="text"
+        [(ngModel)]="newCookbookName"
+        placeholder="Cookbook Name (e.g. Desserts)"
+        class="w-full mb-3 px-4 py-3 rounded-xl bg-stone-50 border border-stone-200 focus:outline-none focus:ring-2 focus:ring-green-500"
+      />
+      <input
+        type="text"
+        [(ngModel)]="newCookbookDesc"
+        placeholder="Description (Optional)"
+        class="w-full mb-6 px-4 py-3 rounded-xl bg-stone-50 border border-stone-200 focus:outline-none focus:ring-2 focus:ring-green-500"
+      />
+      <div class="flex gap-3">
+        <button
+          (click)="close()"
+          class="flex-1 py-3 bg-stone-100 text-stone-600 font-bold rounded-xl hover:bg-stone-200"
+        >
+          Cancel
+        </button>
+        <button
+          (click)="create()"
+          [disabled]="!newCookbookName().trim() || isCreatingCookbook()"
+          class="flex-1 py-3 bg-green-600 text-white font-bold rounded-xl hover:bg-green-700 shadow-lg shadow-green-600/20 disabled:opacity-50 disabled:cursor-not-allowed"
+        >
+          {{ isCreatingCookbook() ? 'Creating…' : 'Create' }}
+        </button>
+      </div>
+    </div>
+  </div>
+}
